+++ austrakka/components/fieldtype/funcs.py
@@ -1,12 +1,10 @@
 from typing import List
 
-import pandas as pd
-
 from austrakka.utils.api import api_get
 from austrakka.utils.api import api_post
 from austrakka.utils.api import api_put
 from austrakka.utils.misc import logger_wraps
-from austrakka.utils.output import print_formatted
+from austrakka.utils.output import print_dict
 from austrakka.utils.paths import METADATACOLUMNTYPE_PATH
 from austrakka.utils.helpers.fieldtype import get_fieldtype_by_name
 
@@ -21,12 +19,7 @@ def list_fieldtypes(out_format: str):
     for row in data:
         if 'validValues' in row:
             row['validValues'] = ",".join([val['name'] for val in row['validValues']])
-    result = pd.DataFrame.from_dict(data)
-
-    print_formatted(
-        result,
-        out_format,
-    )
+    print_dict(data, out_format)
 
 
 @logger_wraps()