+++ austrakka/components/proforma/__init__.py
@@ -13,7 +13,10 @@ from .funcs import \
     enable_proforma, \
     share_proforma, \
     unshare_proforma, \
-    list_groups_proforma, attach_proforma, pull_proforma
+    list_groups_proforma, \
+    attach_proforma, \
+    generate_proforma, \
+    pull_proforma
 
 from ...utils.options import *
 
@@ -129,6 +132,26 @@ def proforma_attach(abbrev: str,
     else:
         attach_proforma(abbrev, file_path)
 
+@proforma.command('generate')
+@click.argument('abbrev', type=click.STRING)
+@click.option('-r', '--restrict', type=click.STRING, multiple=True, nargs=2,
+    help='Key-value pair; restrict the specified field to the specified '+
+        'comma-separated subset of values')
+@click.option(
+    '--nndss/--no-nndss',
+    type=bool,
+    default=False,
+    help='Include an NNDSS label column in the Data Dictionary',
+)
+@opt_project(
+    required=False,
+    help="Project to fill in on the Groups for Sharing tab",
+)
+def proforma_generate(abbrev: str, restrict, nndss, project):
+    """
+    Generate a draft XLSX pro forma template from the current specification.
+    """
+    generate_proforma(abbrev, restrict, nndss_column=nndss, project_abbrev=project)
 
 @proforma.command('list')
 @table_format_option()