+++ austrakka/components/project/__init__.py
@@ -1,6 +1,7 @@
 # pylint: disable=expression-not-assigned
 import click
-from austrakka.utils.output import table_format_option, object_format_option
+from austrakka.utils.output import table_format_option
+from austrakka.utils.output import object_format_option
 from austrakka.utils.cmd_filter import hide_admin_cmds
 from austrakka.utils.options import opt_abbrev, opt_is_active
 from austrakka.utils.options import opt_name