+++ austrakka/components/project/metadata/__init__.py
@@ -1,9 +1,11 @@
+from typing import Optional
+
 import click
 
 from austrakka.utils.output import table_format_option
-from austrakka.utils.options import \
-    opt_merge_algorithm, \
-    opt_output_dir
+from austrakka.utils.output import FORMATS
+from austrakka.utils.options import opt_merge_algorithm
+from austrakka.utils.options import opt_view_id
 
 from .funcs import list_dataset_views, \
     download_dataset_view, \
@@ -34,16 +36,16 @@ def get_dataset_view_list(abbrev: str, out_format: str):
 
 
 @metadata.command('get')
-@click.option('-id',
-              '--view-id',
-              help='Project metadata view to get', )
+@opt_view_id()
 @click.argument('abbrev', type=str)
-@opt_output_dir()
-def get_dataset_view(abbrev: str,
-                     output_dir: str,
-                     view_id: str,):
+@table_format_option(FORMATS.CSV)
+def get_dataset_view(
+        abbrev: str,
+        view_id: Optional[str],
+        out_format: str,
+):
     """Get a specific metadata view."""
-    download_dataset_view(output_dir, view_id, abbrev)
+    download_dataset_view(view_id, abbrev, out_format)
 
 
 @metadata.command('set-merge')